+++ methods/parallel_cpp/kernel.hpp
@@ -17,26 +17,26 @@ struct ModelConfig {
 };
 
 struct LayerWeights {
-  std::vector<double> attn_wq;
-  std::vector<double> attn_wk;
-  std::vector<double> attn_wv;
-  std::vector<double> attn_wo;
-  std::vector<double> mlp_fc1;
-  std::vector<double> mlp_fc2;
+  std::vector<float> attn_wq;
+  std::vector<float> attn_wk;
+  std::vector<float> attn_wv;
+  std::vector<float> attn_wo;
+  std::vector<float> mlp_fc1;
+  std::vector<float> mlp_fc2;
 };
 
 struct Model {
   ModelConfig config;
-  std::vector<double> wte;
-  std::vector<double> wpe;
-  std::vector<double> lm_head;
+  std::vector<float> wte;
+  std::vector<float> wpe;
+  std::vector<float> lm_head;
   std::vector<LayerWeights> layers;
 };
 
 struct KernelResult {
   int seq_len = 0;
-  std::vector<double> logits;
-  double loss = 0.0;
+  std::vector<float> logits;
+  float loss = 0.0f;
 };
 
 struct BatchTokens {
@@ -60,36 +60,36 @@ template <typename T> struct DeviceBuffer {
 
 struct DeviceModel {
   ModelConfig config;
-  DeviceBuffer<double> wte;
-  DeviceBuffer<double> wpe;
-  DeviceBuffer<double> lm_head;
-  std::vector<DeviceBuffer<double>> attn_wq;
-  std::vector<DeviceBuffer<double>> attn_wk;
-  std::vector<DeviceBuffer<double>> attn_wv;
-  std::vector<DeviceBuffer<double>> attn_wo;
-  std::vector<DeviceBuffer<double>> mlp_fc1;
-  std::vector<DeviceBuffer<double>> mlp_fc2;
+  DeviceBuffer<float> wte;
+  DeviceBuffer<float> wpe;
+  DeviceBuffer<float> lm_head;
+  std::vector<DeviceBuffer<float>> attn_wq;
+  std::vector<DeviceBuffer<float>> attn_wk;
+  std::vector<DeviceBuffer<float>> attn_wv;
+  std::vector<DeviceBuffer<float>> attn_wo;
+  std::vector<DeviceBuffer<float>> mlp_fc1;
+  std::vector<DeviceBuffer<float>> mlp_fc2;
 };
 
 struct DeviceWorkspace {
   DeviceBuffer<int> tokens;
-  DeviceBuffer<double> embeddings;
-  DeviceBuffer<double> hidden;
+  DeviceBuffer<float> embeddings;
+  DeviceBuffer<float> hidden;
   // Forward-only transformer workspace and caches for future CUDA kernels.
-  DeviceBuffer<double> x;
-  DeviceBuffer<double> x_tmp;
-  DeviceBuffer<double> x_mid;
-  DeviceBuffer<double> x_norm2;
-  DeviceBuffer<double> norm;
-  DeviceBuffer<double> q;
-  DeviceBuffer<double> k_cache;
-  DeviceBuffer<double> v_cache; 
-  DeviceBuffer<double> attn_out;
-  DeviceBuffer<double> mlp_hidden;
-  DeviceBuffer<double> logits;
-  DeviceBuffer<double> loss;
-  DeviceBuffer<double> relu; 
-  DeviceBuffer<double> fc2; 
+  DeviceBuffer<float> x;
+  DeviceBuffer<float> x_tmp;
+  DeviceBuffer<float> x_mid;
+  DeviceBuffer<float> x_norm2;
+  DeviceBuffer<float> norm;
+  DeviceBuffer<float> q;
+  DeviceBuffer<float> k_cache;
+  DeviceBuffer<float> v_cache; 
+  DeviceBuffer<float> attn_out;
+  DeviceBuffer<float> mlp_hidden;
+  DeviceBuffer<float> logits;
+  DeviceBuffer<float> loss;
+  DeviceBuffer<float> relu; 
+  DeviceBuffer<float> fc2; 
 };
 
 struct KernelLaunch {
@@ -117,7 +117,7 @@ inline KernelLaunch make_1d_launch(std::size_t work_items, int threads = 256) {
 Model make_empty_model(const ModelConfig &config);
 Model initialize_model(const ModelConfig &config, std::uint32_t seed);
 void load_model_from_f32(Model &host_model, const std::vector<float> &values);
-std::vector<double> flatten_model_values(const Model &host_model);
+std::vector<float> flatten_model_values(const Model &host_model);
 DeviceModel upload_model_to_device(const Model &host_model);
 void free_device_model(DeviceModel *device_model);
 KernelResult run_forward_batched(const DeviceModel &device_model,
