+++ methods/serial_cpp/kernel.cpp
@@ -7,27 +7,27 @@
 namespace {
 
 struct LayerStepCache {
-    std::vector<double> k;
-    std::vector<double> v;
+    std::vector<float> k;
+    std::vector<float> v;
 };
 
 struct StepForwardCache {
     std::vector<LayerStepCache> layers;
 };
 
-constexpr double kRmsNormEps = 1e-5;
+constexpr float kRmsNormEps = 1e-5f;
 
-std::vector<double> add_vectors(const std::vector<double>& left, const std::vector<double>& right) {
+std::vector<float> add_vectors(const std::vector<float>& left, const std::vector<float>& right) {
     // Pure elementwise vector add:
     // output[i] = left[i] + right[i]
-    std::vector<double> output(left.size(), 0.0);
+    std::vector<float> output(left.size(), 0.0f);
     for (std::size_t idx = 0; idx < left.size(); ++idx) {
         output[idx] = left[idx] + right[idx];
     }
     return output;
 }
 
-std::vector<double> linear(const std::vector<double>& input, const std::vector<double>& weights, int out_dim) {
+std::vector<float> linear(const std::vector<float>& input, const std::vector<float>& weights, int out_dim) {
     // Dense matrix-vector multiply.
     // Shapes:
     // - input: [in_dim]
@@ -37,9 +37,9 @@ std::vector<double> linear(const std::vector<double>& input, const std::vector<d
     // Math:
     // output[out] = sum_in weights[out, in] * input[in]
     const int in_dim = static_cast<int>(input.size());
-    std::vector<double> output(out_dim, 0.0);
+    std::vector<float> output(out_dim, 0.0f);
     for (int out = 0; out < out_dim; ++out) {
-        double sum = 0.0;
+        float sum = 0.0f;
         for (int in = 0; in < in_dim; ++in) {
             sum += weights[out * in_dim + in] * input[in];
         }
@@ -48,27 +48,27 @@ std::vector<double> linear(const std::vector<double>& input, const std::vector<d
     return output;
 }
 
-std::vector<double> softmax(const std::vector<double>& logits) {
+std::vector<float> softmax(const std::vector<float>& logits) {
     // Convert arbitrary logits into a probability distribution.
     //
     // Math:
     // probs[i] = exp(logits[i]) / sum_j exp(logits[j])
     //
     // We subtract max_logit first for numerical stability so exp() does not overflow.
-    const double max_logit = *std::max_element(logits.begin(), logits.end());
-    std::vector<double> probs(logits.size(), 0.0);
-    double exp_sum = 0.0;
+    const float max_logit = *std::max_element(logits.begin(), logits.end());
+    std::vector<float> probs(logits.size(), 0.0f);
+    float exp_sum = 0.0f;
     for (std::size_t idx = 0; idx < logits.size(); ++idx) {
         probs[idx] = std::exp(logits[idx] - max_logit);
         exp_sum += probs[idx];
     }
-    for (double& prob : probs) {
+    for (float& prob : probs) {
         prob /= exp_sum;
     }
     return probs;
 }
 
-std::vector<double> rmsnorm(const std::vector<double>& input) {
+std::vector<float> rmsnorm(const std::vector<float>& input) {
     // RMSNorm rescales the vector by its root-mean-square magnitude.
     //
     // Math:
@@ -77,14 +77,14 @@ std::vector<double> rmsnorm(const std::vector<double>& input) {
     // output[i] = input[i] * scale
     //
     // This keeps the direction of the vector but normalizes its overall scale.
-    double mean_square = 0.0;
-    for (double value : input) {
+    float mean_square = 0.0f;
+    for (float value : input) {
         mean_square += value * value;
     }
-    mean_square /= static_cast<double>(input.size());
-    const double scale = 1.0 / std::sqrt(mean_square + kRmsNormEps);
+    mean_square /= static_cast<float>(input.size());
+    const float scale = 1.0f / std::sqrt(mean_square + kRmsNormEps);
 
-    std::vector<double> output(input.size(), 0.0);
+    std::vector<float> output(input.size(), 0.0f);
     for (std::size_t idx = 0; idx < input.size(); ++idx) {
         output[idx] = input[idx] * scale;
     }
@@ -106,79 +106,79 @@ KernelResult forward_pass(const Model& model, const std::vector<int>& tokens) {
         StepForwardCache& step_cache = cache[pos];
         const int input_token = tokens[pos];
         const int target_token = tokens[pos + 1];
-        std::vector<double> embed_pre(config.n_embd, 0.0);
+        std::vector<float> embed_pre(config.n_embd, 0.0f);
 
         // Input: one token ID and one position ID.
         // Transformation: look up the token embedding and position embedding, add them elementwise.
         // Output: the pre-normalized representation for this sequence position.
         //* CUDA speedup opporunity //* Per embedding dim
         for (int col = 0; col < config.n_embd; ++col) {
-            double token_val = model.wte[input_token * config.n_embd + col];
-            double pos_val   = model.wpe[pos * config.n_embd + col];
+            float token_val = model.wte[input_token * config.n_embd + col];
+            float pos_val   = model.wpe[pos * config.n_embd + col];
 
             embed_pre[col] = token_val + pos_val;
         }
 
-        std::vector<double> x = rmsnorm(embed_pre); //* CUDA opporunity
+        std::vector<float> x = rmsnorm(embed_pre); //* CUDA opporunity
         step_cache.layers.resize(config.n_layer);
 
         for (int layer_idx = 0; layer_idx < config.n_layer; ++layer_idx) {
             const LayerWeights& layer = model.layers[layer_idx];
             LayerStepCache& layer_step = step_cache.layers[layer_idx];
-            const std::vector<double> x_residual = x;
-            const std::vector<double> x_norm1 = rmsnorm(x);
-            const std::vector<double> q = linear(x_norm1, layer.attn_wq, config.n_embd);
+            const std::vector<float> x_residual = x;
+            const std::vector<float> x_norm1 = rmsnorm(x);
+            const std::vector<float> q = linear(x_norm1, layer.attn_wq, config.n_embd);
             layer_step.k = linear(x_norm1, layer.attn_wk, config.n_embd);
             layer_step.v = linear(x_norm1, layer.attn_wv, config.n_embd);
 
-            std::vector<double> x_attn(config.n_embd, 0.0);
+            std::vector<float> x_attn(config.n_embd, 0.0f);
             for (int head = 0; head < config.n_head; ++head) {
                 const int head_start = head * config.head_dim();
-                std::vector<double> attn_logits(pos + 1, 0.0);
+                std::vector<float> attn_logits(pos + 1, 0.0f);
                 for (int t = 0; t <= pos; ++t) {
-                    const std::vector<double>& past_k = cache[t].layers[layer_idx].k;
-                    double dot = 0.0;
+                    const std::vector<float>& past_k = cache[t].layers[layer_idx].k;
+                    float dot = 0.0f;
                     for (int j = 0; j < config.head_dim(); ++j) {
                         dot += q[head_start + j] * past_k[head_start + j];
                     }
-                    attn_logits[t] = dot / std::sqrt(static_cast<double>(config.head_dim()));
+                    attn_logits[t] = dot / std::sqrt(static_cast<float>(config.head_dim()));
                 }
 
-                const std::vector<double> attn_weights = softmax(attn_logits);
+                const std::vector<float> attn_weights = softmax(attn_logits);
                 for (int t = 0; t <= pos; ++t) {
-                    const std::vector<double>& past_v = cache[t].layers[layer_idx].v;
-                    const double weight = attn_weights[t];
+                    const std::vector<float>& past_v = cache[t].layers[layer_idx].v;
+                    const float weight = attn_weights[t];
                     for (int j = 0; j < config.head_dim(); ++j) {
                         x_attn[head_start + j] += weight * past_v[head_start + j];
                     }
                 }
             }
 
-            const std::vector<double> attn_proj = linear(x_attn, layer.attn_wo, config.n_embd);
-            const std::vector<double> x_mid = add_vectors(x_residual, attn_proj);
-            const std::vector<double> x_norm2 = rmsnorm(x_mid);
-            std::vector<double> relu = linear(x_norm2, layer.mlp_fc1, config.mlp_dim());
-            for (double& value : relu) {
-                value = std::max(0.0, value);
+            const std::vector<float> attn_proj = linear(x_attn, layer.attn_wo, config.n_embd);
+            const std::vector<float> x_mid = add_vectors(x_residual, attn_proj);
+            const std::vector<float> x_norm2 = rmsnorm(x_mid);
+            std::vector<float> relu = linear(x_norm2, layer.mlp_fc1, config.mlp_dim());
+            for (float& value : relu) {
+                value = std::max(0.0f, value);
             }
 
             // Input: the post-attention hidden state.
             // Transformation: expand with FC1, apply ReLU, project back down with FC2, then add the residual.
             // Output: the layer output that becomes the next layer's input.
-            const std::vector<double> fc2 = linear(relu, layer.mlp_fc2, config.n_embd);
+            const std::vector<float> fc2 = linear(relu, layer.mlp_fc2, config.n_embd);
             x = add_vectors(x_mid, fc2);
         }
 
         // Input: the final hidden state at this position.
         // Transformation: project into vocabulary space and normalize with softmax for the target token.
         // Output: logits for every possible next token and one scalar loss contribution.
-        const std::vector<double> logits = linear(x, model.lm_head, config.vocab_size);
-        const std::vector<double> probs = softmax(logits);
+        const std::vector<float> logits = linear(x, model.lm_head, config.vocab_size);
+        const std::vector<float> probs = softmax(logits);
         result.logits.insert(result.logits.end(), logits.begin(), logits.end());
         result.loss += -std::log(probs[target_token]);
     }
 
-    result.loss /= static_cast<double>(result.seq_len);
+    result.loss /= static_cast<float>(result.seq_len);
     return result;
 }
